Move Images
Move images to src/images
