+++ src/orgs/smartsaver/sections/survey/Banner.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Section from './Section'
-import logo from '../../images/banner.jpg'
+import logo from '../../../../images/smartsaver/maeghan-david-banner.jpg'
 import style from './Banner.module.css'
 import { introContent } from '../../content/survey/introContent'
