+++ src/orgs/smartsaver/sections/ufileresp/UfilerespLayout.js
@@ -2,7 +2,7 @@ import React from 'react'
 import Layout from '../../../../components/Layout'
 import Navbar from '../../../../components/Navbar'
 import NavbarItem from '../../../../components/Navbar/NavbarItem'
-import logo from '../../images/ufile-logo.png'
+import logo from '../../../../images/smartsaver/ufile-logo.png'
 
 const UfilerespLayout = ({ children }) => {
   return (
